+++ packages/office-ui-fabric-react/src/components/ComboBox/ComboBox.types.ts
@@ -5,6 +5,7 @@ import { IStyle, ITheme } from '../../Styling';
 import { IButtonStyles } from '../../Button';
 import { IRenderFunction } from '../../Utilities';
 import { IComboBoxClassNames } from './ComboBox.classNames';
+import { IKeytipProps } from '../../Keytip';
 
 export interface IComboBox {
   /**
@@ -97,7 +98,7 @@ export interface IComboBoxProps extends ISelectableDroppableTextProps<IComboBox>
   /**
    * Value to show in the input, does not have to map to a combobox option
    */
-  value?: string;
+  text?: string;
 
   /**
    * The IconProps to use for the button aspect of the combobox
@@ -171,6 +172,17 @@ export interface IComboBoxProps extends ISelectableDroppableTextProps<IComboBox>
    * @default true
    */
   isButtonAriaHidden?: boolean;
+
+  /**
+   * Optional keytip for this combo box
+   */
+  keytipProps?: IKeytipProps;
+
+  /**
+   * Value to show in the input, does not have to map to a combobox option
+   * @deprecated Use `text` instead.
+   */
+  value?: string;
 }
 
 export interface IComboBoxStyles {