+++ packages/office-ui-fabric-react/src/components/ContextualMenu/ContextualMenu.classNames.ts
@@ -1,8 +1,8 @@
-import { memoizeFunction } from '../../Utilities';
-import { ITheme, mergeStyleSets } from '../../Styling';
-import { getStyles as getContextualMenuStyles, getMenuItemStyles } from './ContextualMenu.styles';
-import { IVerticalDividerClassNames } from '../Divider/VerticalDivider.types';
 import { getDividerClassNames } from '../Divider/VerticalDivider.classNames';
+import { getMenuItemStyles, getStyles as getContextualMenuStyles } from './ContextualMenu.styles';
+import { ITheme, mergeStyleSets } from '../../Styling';
+import { IVerticalDividerClassNames } from '../Divider/VerticalDivider.types';
+import { memoizeFunction } from '../../Utilities';
 
 export interface IContextualMenuClassNames {
   container: string;
@@ -21,6 +21,7 @@ export interface IMenuItemClassNames {
   checkmarkIcon: string;
   subMenuIcon: string;
   label: string;
+  secondaryText: string;
   splitContainer: string;
   splitPrimary: string;
   splitMenu: string;
@@ -81,6 +82,7 @@ export const getItemClassNames = memoizeFunction((
   dividerClassName?: string,
   iconClassName?: string,
   subMenuClassName?: string,
+  primaryDisabled?: boolean
 ): IMenuItemClassNames => {
 
   const styles = getMenuItemStyles(theme);
@@ -127,11 +129,11 @@ export const getItemClassNames = memoizeFunction((
         'is-checked',
         styles.rootChecked
       ],
-      disabled && [
+      (disabled || primaryDisabled) && [
         'is-disabled',
         styles.rootDisabled
       ],
-      !disabled && !checked && [{
+      !(disabled || primaryDisabled) && !checked && [{
         selectors: {
           ':hover': styles.rootHovered,
           ':active': styles.rootPressed,
@@ -197,6 +199,10 @@ export const getItemClassNames = memoizeFunction((
     label: [
       'ms-ContextualMenu-itemText',
       styles.label
+    ],
+    secondaryText: [
+      'ms-ContextualMenu-secondaryText',
+      styles.secondaryText
     ],
     splitContainer: [
       styles.splitButtonFlexContainer,