+++ packages/office-ui-fabric-react/src/components/ScrollablePane/ScrollablePane.styles.ts
@@ -1,31 +1,57 @@
 import { IScrollablePaneStyleProps, IScrollablePaneStyles } from './ScrollablePane.types';
 import {
   HighContrastSelector,
-  IStyle
+  IStyle,
+  ZIndexes,
+  getGlobalClassNames,
 } from '../../Styling';
+
+const GlobalClassNames = {
+  root: 'ms-ScrollablePane',
+  contentContainer: 'ms-ScrollablePane--contentContainer'
+};
 
 export const getStyles = (
   props: IScrollablePaneStyleProps
 ): IScrollablePaneStyles => {
-  const { className } = props;
+  const {
+    className,
+    theme,
+  } = props;
+
+  const classNames = getGlobalClassNames(GlobalClassNames, theme);
 
   const AboveAndBelowStyles: IStyle = {
     position: 'absolute',
     pointerEvents: 'auto',
     width: '100%',
-    zIndex: 1
+    zIndex: ZIndexes.ScrollablePane,
+    overflowY: 'hidden',
+    overflowX: 'auto'
+  };
+
+  const positioningStyle: IStyle = {
+    zIndex: ZIndexes.ScrollablePane,
+    position: 'absolute',
+    top: 0,
+    right: 0,
+    bottom: 0,
+    left: 0,
+    WebkitOverflowScrolling: 'touch'
   };
 
   return ({
     root: [
-      'ms-ScrollablePane',
-      {
-        overflowY: 'auto',
-        maxHeight: 'inherit',
-        height: 'inherit',
-        WebkitOverflowScrolling: 'touch'
-      },
+      classNames.root,
+      positioningStyle,
       className
+    ],
+    contentContainer: [
+      classNames.contentContainer,
+      {
+        overflowY: 'auto'
+      },
+      positioningStyle
     ],
     stickyAbove: [
       {
@@ -46,6 +72,12 @@ export const getStyles = (
             borderTop: '1px solid WindowText'
           }
         }
+      },
+      AboveAndBelowStyles
+    ],
+    stickyBelowItems: [
+      {
+        bottom: 0
       },
       AboveAndBelowStyles
     ]